fixed problem with ivy and ^M in line endings
When files contain lines which end with ^m^l the changes are not saved. The
source of the problem is based in the fact, that in the original file buffer
^m^l is understood as line ending, while in the grep buffer only ^l is the line
ending and ^m is a seperate character. The result is, that the check between
original file and buffer fails, and thus the content won't be saved.

To fix this a suffix of ^m in the old _and_ new string is ignored. The later one
is important, as otherwise there would be ^m^m in the file after the replacement.